Initialized update_check after settings (#731)
* Re-factored update check

* Important dragonfly after initialize functions

* dep_missing first

* Utilized printer.out

* Simplified update_timer

* Removed `"python", "-m"`, from update check
