+++ _caster.py
@@ -5,16 +5,22 @@ Created on Jun 29, 2014
 '''
 import datetime
 import logging
-from dragonfly import get_engine
-from dragonfly.windows.window import Window
 
 logging.basicConfig(format = "%(asctime)s : %(levelname)s : %(funcName)s\n%(msg)s")   
 
-from castervoice.lib import settings  # requires toml
-settings.initialize()
-
 from castervoice.lib.ctrl.dependencies import DependencyMan  # requires nothing
 DependencyMan().initialize()
+
+from castervoice.lib import settings  
+settings.initialize()
+
+from castervoice.lib.ctrl.updatecheck import UpdateChecker # requires settings/dependencies
+UpdateChecker().initialize()
+
+from dragonfly import get_engine
+from dragonfly.windows.window import Window
+
+
 _NEXUS = None
 
 class LoggingHandler(logging.Handler):
