+++ castervoice/rules/core/utility_rules/caster_rule.py
@@ -1,7 +1,8 @@
 from dragonfly import MappingRule, Function, RunCommand, Playback
 
 from castervoice.lib import control
-from castervoice.lib.ctrl.dependencies import update, find_pip
+from castervoice.lib.ctrl.dependencies import find_pip
+from castervoice.lib.ctrl.updatecheck import update
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
 from castervoice.lib.merge.state.short import R
